added better reservation management

+++ backend/src/main/java/finki/db/tasty_tabs/service/impl/ReservationServiceImpl.java
@@ -130,7 +130,7 @@ public class ReservationServiceImpl implements ReservationService {
         ReservationManagedFrontStaffId id = new ReservationManagedFrontStaffId(
                 reservation.getId(),
                 frontStaff.getId(),
-                table.getTableNumber().longValue()
+                table.getTableNumber()
         );
 
         ReservationManagedFrontStaff managedReservation = new ReservationManagedFrontStaff();
